refactor!: move loader API interfaces, and more
