--- allium/src/main/java/dev/hugeblank/allium/loader/LibraryInitializer.java
@@ -1,7 +1,0 @@
-package dev.hugeblank.allium.loader;
-
-import dev.hugeblank.allium.loader.type.WrappedLuaLibrary;
-
-public interface LibraryInitializer {
-    WrappedLuaLibrary init(Script script);
-}
