--- allium/src/main/java/dev/hugeblank/allium/loader/ScriptResource.java
@@ -1,6 +1,0 @@
-package dev.hugeblank.allium.loader;
-
-public interface ScriptResource extends AutoCloseable {
-    @Override
-    void close() throws Exception;
-}
