--- allium/src/main/java/dev/hugeblank/allium/loader/type/AlliumExtension.java
@@ -1,10 +1,0 @@
-package dev.hugeblank.allium.loader.type;
-
-import dev.hugeblank.allium.loader.ScriptExecutor;
-
-import java.util.function.Consumer;
-
-@SuppressWarnings("ALL")
-public interface AlliumExtension {
-    void onInitialize();
-}
