+++ allium/src/main/java/dev/hugeblank/allium/loader/type/StaticBinder.java
@@ -1,7 +1,10 @@
 package dev.hugeblank.allium.loader.type;
 
 import dev.hugeblank.allium.Allium;
+import dev.hugeblank.allium.loader.type.coercion.TypeCoercions;
 import dev.hugeblank.allium.loader.type.property.PropertyData;
+import dev.hugeblank.allium.util.AnnotationUtils;
+import dev.hugeblank.allium.util.ArgumentUtils;
 import dev.hugeblank.allium.util.JavaHelpers;
 import me.basiqueevangelist.enhancedreflection.api.EClass;
 import me.basiqueevangelist.enhancedreflection.api.EMember;
