+++ allium/src/main/java/dev/hugeblank/allium/loader/type/UDFFunctions.java
@@ -1,5 +1,7 @@
 package dev.hugeblank.allium.loader.type;
 
+import dev.hugeblank.allium.loader.type.coercion.TypeCoercions;
+import dev.hugeblank.allium.util.ArgumentUtils;
 import me.basiqueevangelist.enhancedreflection.api.EClass;
 import me.basiqueevangelist.enhancedreflection.api.EMethod;
 import me.basiqueevangelist.enhancedreflection.api.typeuse.EClassUse;
