+++ allium/src/main/java/dev/hugeblank/allium/loader/type/property/PropertyMethodData.java
@@ -3,8 +3,8 @@ package dev.hugeblank.allium.loader.type.property;
 import dev.hugeblank.allium.loader.type.InvalidArgumentException;
 import me.basiqueevangelist.enhancedreflection.api.EMethod;
 import me.basiqueevangelist.enhancedreflection.api.typeuse.EClassUse;
-import dev.hugeblank.allium.loader.type.ArgumentUtils;
-import dev.hugeblank.allium.loader.type.TypeCoercions;
+import dev.hugeblank.allium.util.ArgumentUtils;
+import dev.hugeblank.allium.loader.type.coercion.TypeCoercions;
 import org.jetbrains.annotations.Nullable;
 import org.squiddev.cobalt.Constants;
 import org.squiddev.cobalt.LuaError;
