+++ allium/src/main/java/dev/hugeblank/allium/loader/type/property/PropertyResolver.java
@@ -1,7 +1,7 @@
 package dev.hugeblank.allium.loader.type.property;
 
 import dev.hugeblank.allium.Allium;
-import dev.hugeblank.allium.loader.type.AnnotationUtils;
+import dev.hugeblank.allium.util.AnnotationUtils;
 import me.basiqueevangelist.enhancedreflection.api.EClass;
 import me.basiqueevangelist.enhancedreflection.api.EField;
 import me.basiqueevangelist.enhancedreflection.api.EMethod;
