+++ bouquet/src/main/java/dev/hugeblank/bouquet/api/lib/DefaultEventsLib.java
@@ -1,6 +1,6 @@
 package dev.hugeblank.bouquet.api.lib;
 
-import dev.hugeblank.allium.loader.type.WrappedLuaLibrary;
+import dev.hugeblank.allium.api.WrappedLuaLibrary;
 import dev.hugeblank.bouquet.api.event.CommonEventHandlers;
 import dev.hugeblank.bouquet.api.event.SimpleEventType;
 import dev.hugeblank.allium.loader.type.annotation.LuaWrapped;
