+++ bouquet/src/main/java/dev/hugeblank/bouquet/api/lib/fs/LuaHandle.java
@@ -1,7 +1,7 @@
 package dev.hugeblank.bouquet.api.lib.fs;
 
 import dev.hugeblank.allium.loader.Script;
-import dev.hugeblank.allium.loader.ScriptResource;
+import dev.hugeblank.allium.api.ScriptResource;
 import org.squiddev.cobalt.LuaError;
 
 import java.io.Closeable;
